changed to iterative merkle tree generation, rather than recursive

+++ block/block.go
@@ -10,35 +10,34 @@ import (
 )
 
 type Block struct {
-	Id      int
-	Hash    Hash
-	Trans   []Transaction
-	Sig     Signature
+	Id    int
+	Hash  Hash
+	Trans []Transaction
+	Sig   Signature
 }
 
 type Hash struct {
-	Hash    []byte // hash of previous block
-	Proof   PoS
+	Hash  []byte // hash of previous block
+	Proof PoS
 }
 
 type PoS struct {
-	Commit          pos.Commitment
-	Challenge       []byte   // this round's challenge
-	Answer          Answer   // answer to the challenge and proof
-	Quality         float64  // quality of the answer
+	Commit    pos.Commitment
+	Challenge []byte  // this round's challenge
+	Answer    Answer  // answer to the challenge and proof
+	Quality   float64 // quality of the answer
 }
 
 type Answer struct {
-	Size    int
-	Hashes  [][]byte
-	Proofs  [][][]byte
+	Size   int64
+	Hashes [][]byte
+	Proofs [][][]byte
 }
 
 type Signature struct {
-	Tsig    []byte // signature on transaction i
-	Ssig    []byte // signature on signature i-1
+	Tsig []byte // signature on transaction i
+	Ssig []byte // signature on signature i-1
 }
-
 
 func NewBlock(old *Block, prf PoS, ts []Transaction, signer crypto.Signer) *Block {
 	oldH, err := old.Hash.MarshalBinary()
@@ -47,7 +46,7 @@ func NewBlock(old *Block, prf PoS, ts []Transaction, signer crypto.Signer) *Bloc
 	}
 	prevHash := sha3.Sum256(oldH)
 	h := Hash{
-		Hash: prevHash[:],
+		Hash:  prevHash[:],
 		Proof: prf,
 	}
 
@@ -57,7 +56,7 @@ func NewBlock(old *Block, prf PoS, ts []Transaction, signer crypto.Signer) *Bloc
 		if err != nil {
 			panic(err)
 		}
-		tsBytes = append(tsBytes, b ...)
+		tsBytes = append(tsBytes, b...)
 	}
 	sigBytes := util.Concat([][]byte{old.Sig.Tsig, old.Sig.Ssig})
 
@@ -75,14 +74,13 @@ func NewBlock(old *Block, prf PoS, ts []Transaction, signer crypto.Signer) *Bloc
 	}
 
 	b := Block{
-		Id: old.Id + 1,
-		Hash: h,
+		Id:    old.Id + 1,
+		Hash:  h,
 		Trans: ts,
-		Sig: sig,
+		Sig:   sig,
 	}
 	return &b
 }
-
 
 func (b *Block) MarshalBinary() ([]byte, error) {
 	return json.Marshal(b)
